+++ CHANGELOG.md
@@ -6,6 +6,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Use correct tag for media:player
+
+## [0.1.1] - 2022-12-06
+
+### Fixed
+- Require released version for feed generator
+
+## [0.1.0] - 2022-12-05
+
 First preview release
 
+[Unreleased]: https://github.com/brotkrueml/typo3-feed-generator/compare/v0.1.1...HEAD
+[0.1.1]: https://github.com/brotkrueml/typo3-feed-generator/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/brotkrueml/typo3-feed-generator-mrss/releases/tag/v0.1.0